+++ src/main/java/com/playus/twpservice/domain/party/dto/detail/PartyDetailResponse.java
@@ -1,4 +1,4 @@
-package com.playus.twpservice.domain.party.dto.partydescription;
+package com.playus.twpservice.domain.party.dto.detail;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.playus.twpservice.domain.party.enums.PartyAgeGroup;
