+++ src/main/java/com/playus/twpservice/domain/party/entity/Party.java
@@ -1,7 +1,7 @@
 package com.playus.twpservice.domain.party.entity;
 
 import com.playus.twpservice.domain.common.data.BaseTimeEntity;
-import com.playus.twpservice.domain.party.dto.partyupdate.PartyUpdateRequest;
+import com.playus.twpservice.domain.party.dto.update.PartyUpdateRequest;
 import com.playus.twpservice.domain.party.enums.PartyJoinMethod;
 import com.playus.twpservice.domain.party.enums.PartyGender;
 import jakarta.persistence.*;
